refactor navBar to use new navModule component
