+++ packages/player/src/components/navbar/_navbar.scss
@@ -38,6 +38,7 @@
       border: none;
       display: flex;
       align-items: center;
+      margin-bottom: 0.25rem;
 
       span {
         transform: rotate(-90deg);
